pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "difftests"
version = "0.1.0"
description = "AI-powered test generator from git diffs — write tests for what you changed"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "faw21" }
]
keywords = ["ai", "testing", "git", "developer-tools", "pytest", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "gitpython>=3.1",
    "openai>=1.0",
    "anthropic>=0.20",
    "pyperclip>=1.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "python-dotenv>=1.0",
]

[project.scripts]
difftests = "difftests.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/difftests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=difftests --cov-report=term-missing"

[tool.coverage.run]
source = ["difftests"]
